requirements.txt
google-generativeai>=0.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.35.0
pandas>=2.0.0

